Separate publisher expectations over several files
