+++ Sources/Circumspect/Diff.swift
@@ -1,0 +1,17 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import Difference
+
+enum AssertDescription {
+    static func difference<T, U>(expected: [T], actual: [U]) -> String {
+        "Expected: \(expected), actual: \(actual)"
+    }
+
+    static func difference<T>(expected: [T], actual: [T]) -> String {
+        diff(expected, actual).joined(separator: ", ")
+    }
+}
